% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_dim.R
\name{desc.dim}
\alias{desc.dim}
\title{Description dimension of MCA analysis}
\usage{
desc.dim(res, nb_dim)
}
\arguments{
\item{res}{results of MCA analysis}

\item{nb_dim}{an integer, number of dimension}
}
\value{
list of description's table
}
\description{
This function returns table of dimension description form MCA analysis
}
