% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_evaluation.R
\name{evaluation}
\alias{evaluation}
\title{Compute Rand index and V measure between real and predicted clusters}
\usage{
evaluation(true_label, pred_label)
}
\arguments{
\item{true_label}{a factor vector of real clusters}

\item{pred_label}{a factor vector of predicted clusters}
}
\value{
Rand index and V measure
}
\description{
This function returns Rand index and V measure between real clusters and predicted clusters
}
\examples{
true_label <-as.factor(c("1","2","2","1"))
pred_label <-as.factor(c("1","2","1","2"))
evaluation(true_label,pred_label)

}
