\name{imcr-package}
\alias{imcr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Is My Cluster Relevant ? IMCR
}

\description{
The method proposed in this package are exploratory multivariate and univariate methods to help interpretation of clusters.
}

\details{
For a complete list of functions, use library(help = "imcr").
}

\author{
Kassandra D'Orazio, Florian Moret, Tanguy Delzant
}

