% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_multivariate_qualitative.R
\name{multi.quali}
\alias{multi.quali}
\title{Multivarial qualitatives variable analysis to describe clusters}
\usage{
multi.quali(
  active_variables,
  clusters,
  quanti.supp = NULL,
  axes = c(1, 2),
  show_graph = NULL
)
}
\arguments{
\item{active_variables}{a data frame. Must contains qualitatives variables}

\item{clusters}{a numeric vector. Must contains clusters from a clustering}

\item{quanti.supp}{index of supplementary variables in the data frame. Muste be character}

\item{axes}{choice of dimensions for graph}

\item{show_graph}{logical. If FALSE, doesn't return any graph}
}
\value{
List of results (table for individuals and variables, correlation between dimension and clusters)
}
\description{
This function realize multivarial analysis to describe clusters whith qualitatives variables
}
