% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_univariate_qualitative.R
\name{plot.uni.quali}
\alias{plot.uni.quali}
\title{Graphics for all the functions of qualitative variable}
\usage{
\method{plot}{uni.quali}(x, type = NULL, name = NULL, digits = 5, ...)
}
\arguments{
\item{x}{a data frame. Contains all the data which needs to be print}

\item{type}{a string. Choose the graphics to be print}

\item{name}{a string. Name of the variable to be print}

\item{digits}{a number. Number of digits for the values}

\item{...}{further arguments of print method}
}
\value{
graphics for the univariate qualitative function
}
\description{
This function allows to print any graphics for the variable univariate qualitative. We override the plot function.
}
