% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_multivariate_quantitative.R
\name{print.multi.quanti}
\alias{print.multi.quanti}
\title{Description of multi.quanti function}
\usage{
\method{print}{multi.quanti}(x, file = NULL, sep = ";", ...)
}
\arguments{
\item{x}{an object of class multi.quanti}

\item{file}{name of file to exports results}

\item{sep}{separator to use in the file}

\item{...}{further arguments of print method}
}
\value{
Description of multi.quanti object
}
\description{
This function print the names and descriptions of multi.quanti function's results
}
