% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_univariate_quantitative.R
\name{print.uni.quanti}
\alias{print.uni.quanti}
\title{Modify print for uni.quanti class}
\usage{
\method{print}{uni.quanti}(x, file = NULL, sep = ";", ...)
}
\arguments{
\item{x}{an object containing the output of one of the univariate quantitative function}

\item{file}{a string. NULL by default, name for the file to export results}

\item{sep}{a string. ";" by default. Separtor for the file}

\item{...}{parameters passed to other methods
#'}
}
\value{

}
\description{
This function overload the print function to facilitate printing several attributes frome the univariate quantitative function suite
}
