% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_value.R
\name{test_value}
\alias{test_value}
\title{Compute test value of quantitative variables and clusters}
\usage{
test_value(active_variables, clusters, show_graph = TRUE, digits = 3)
}
\arguments{
\item{active_variables}{a numeric vector, data frame or matrix}

\item{clusters}{a vector of same length as variables containing the clustering values}

\item{show_graph}{logical. If FALSE, doesn't return the graph}

\item{digits}{a number. Number of digits displayed in tables}
}
\value{
The test value table (returned in console and Viewer tab in Rstudio in html format) and a barplot with means, standard deviation and lines for test value for each variable and for each cluster.
}
\description{
This function returns the test value from quantitative variables and clusters
}
\examples{
data(iris)
iris.cr <- scale(iris[,-5],center=TRUE,scale=TRUE)
d.iris <- dist(iris.cr)
cah.ward <- hclust(d.iris,method="ward.D2")
groupes.cah <- cutree(cah.ward,k=4)
test_value(iris[,-5], groupes.cah, show_graph = TRUE, digits=3)

}
