% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_univariate_qualitative.R
\name{uni.quali}
\alias{uni.quali}
\title{Constructor for qualitative univariate characterization}
\usage{
uni.quali(active_variables, clusters, show_graph = FALSE, digits = 5)
}
\arguments{
\item{active_variables}{a factor vector or matrix. Must contains qualitatives variables}

\item{clusters}{a numeric vector. Must contains clusters from a clustering}

\item{show_graph}{logical. If FALSE, doesn't return any graph}

\item{digits}{a number. Number of digits to display}
}
\value{
List including the result of each function for qualitative univariate characterization
}
\description{
Returns a list of element which allows to characterize univariate qualitative variable
}
\examples{
var = c(rep("yes",7), rep("no",7))
clust = c(1,1,2,1,2,3,1,2,3,3,2,1,3,2)
uni.quali(var,clust)

}
